remove repetitive util functions
