+++ neural_compressor/torch/algorithms/smooth_quant/smooth_quant.py
@@ -17,26 +17,28 @@
 
 import copy
 import json
-
-from neural_compressor.torch.utils import get_ipex_version, logger
+import torch
 
 try:
     import intel_extension_for_pytorch as ipex
 except:
     assert False, "Please install IPEX for smooth quantization."
 
-import torch
 from packaging.version import Version
 
+from neural_compressor.torch.utils import (
+    get_ipex_version, 
+    logger, 
+    dump_model_op_stats, 
+    simple_inference, 
+    ipex_config_path
+)
 from .utility import (
     SQLinearWrapper,
     TorchSmoothQuant,
     cfg_to_qconfig,
-    dump_model_op_stats,
     get_module,
     get_quantizable_ops_recursively,
-    ipex_config_path,
-    simple_inference,
     update_sq_scale,
 )
