+++ requirements_pt.txt
@@ -1,4 +1,5 @@
 intel_extension_for_pytorch
+peft
 py-cpuinfo
 pydantic
 torch
